--- riot_api/__init__.py
@@ -1,4 +1,0 @@
-from .riot_api import API_RIOT, API_LEAGUE, API_VALORANT
-from .riot_api import QueueType
-from .riot_api import Summoner, Riot_Account
-from .riot_api import League_Match
