+++ riot_api_manipulation/internal_helpers.py
@@ -1,0 +1,5 @@
+import json as json_lib
+
+
+def json_format_str(json):
+    return json_lib.dumps(json, indent=2)
